refactor(layout): move codegen symlinks under src/ (part 02/08)
Updates findGeneratedPath in libutil/finder.js so fit-codegen creates
per-package symlinks at libraries/<pkg>/src/generated pointing at the
(unchanged) monorepo-root generated/ directory. Adds mkdir -p for the
target parent so the symlink can be created before src/ exists.

Moves librpc (7 files) and libtype (1 file) root sources into src/.
Both packages' internal ./generated/... imports resolve unchanged
because the symlinks move with the importing files.

Adds an explicit exports map to librpc so the subpath imports used
by generated client files (@forwardimpact/librpc/client.js and
./base.js) keep resolving after the move. Simplifies the dynamic
import in products/guide/lib/status.js to destructure clients from
the main package export rather than reaching into the generated
subpath.

Updates finder.test.js assertions to match the new symlink target.

Part 02 of 08 for spec 390.
