+++ libraries/librpc/src/auth.js
